Added some cargo proxies. Moved CargoBoundsProxy to match the new cargo proxies.
