+++ CCL.Creator/Editor/DistanceDrawer.cs
@@ -1,9 +1,4 @@
 ﻿using CCL.Types.Proxies;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using UnityEditor;
 using UnityEngine;
