+++ CCL.Importer/Proxies/MiscReplacer.cs
@@ -9,7 +9,6 @@ namespace CCL.Importer.Proxies
 {
     [ProxyMap(typeof(TeleportArcPassThroughProxy), typeof(TeleportArcPassThrough))]
     [ProxyMap(typeof(InternalExternalSnapshotSwitcherProxy), typeof(InternalExternalSnapshotSwitcher))]
-    [ProxyMap(typeof(CargoBoundsProxy), typeof(CargoBounds))]
     [Export(typeof(IProxyReplacer))]
     internal class MiscReplacer : Profile, IProxyReplacer
     {
