+++ CCL.Types/Proxies/Cargo/CargoWaterDamageProxy.cs
@@ -1,0 +1,9 @@
+﻿using UnityEngine;
+
+namespace CCL.Types.Proxies.Cargo
+{
+    public class CargoWaterDamageProxy : MonoBehaviour
+    {
+
+    }
+}
